+++ Qtica/__init__.py
@@ -1,6 +1,6 @@
 # Qtica
-__version__ = "0.3.3"
-__version_info__ = (0, 3, 3, "", "")
+__version__ = "0.4.0"
+__version_info__ = (0, 4, 0, "", "")
 
 # Python
 __minimum_python_version__ = (3, 10)