+++ Qtica/core/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 
 from .api import Api
-from ._base import AbstractBase
+from ._base import AbstractBase, Args, MArgs, Func
 from ._qobject import AbstractQObject, QObjectDec
 from ._widget import AbstractWidget, WidgetDec
 from ._tool import AbstractTool, ToolDec
@@ -15,3 +15,6 @@ from ._declarative import (
     BehaviorDec,
     AbstractDec
 )
+from .routes import Routes
+from .qstyle_sheet import QStyleSheet
+from .exception_handler import TryExc