+++ Qtica/widgets/stack.py
@@ -3,7 +3,7 @@ from enum import Enum
 from typing import Union
 from PySide6.QtWidgets import QFrame, QWidget, QLayout, QLayoutItem, QStackedLayout
 from ..utils.alignment import Alignment
-from ..utils.routes import Routes
+from ..core.routes import Routes
 from ..core import AbstractWidget